refactor:создать пакет common,добавить подпакеты controller, service, repository в каждую сущность
