+++ src/main/java/ru/practicum/shareit/common/exception/ConflictException.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.exception;
+package ru.practicum.shareit.common.exception;
 
 public class ConflictException extends RuntimeException {
