+++ src/main/java/ru/practicum/shareit/common/exception/ValidationException.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.exception;
+package ru.practicum.shareit.common.exception;
 
 public class ValidationException extends RuntimeException {
