+++ src/main/java/ru/practicum/shareit/common/handler/ErrorHandler.java
@@ -1,13 +1,13 @@
-package ru.practicum.shareit.handler;
+package ru.practicum.shareit.common.handler;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import ru.practicum.shareit.exception.ConflictException;
-import ru.practicum.shareit.exception.DataNotFoundException;
-import ru.practicum.shareit.exception.ValidationException;
+import ru.practicum.shareit.common.exception.ConflictException;
+import ru.practicum.shareit.common.exception.DataNotFoundException;
+import ru.practicum.shareit.common.exception.ValidationException;
 
 @RestControllerAdvice
 public class ErrorHandler {
