+++ src/main/java/ru/practicum/shareit/item/repository/ItemStorage.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.item;
+package ru.practicum.shareit.item.repository;
 
 import ru.practicum.shareit.item.model.Item;
