+++ src/main/java/ru/practicum/shareit/user/repository/UserStorage.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.user;
+package ru.practicum.shareit.user.repository;
 
 import ru.practicum.shareit.user.model.User;
